refactor(deploy): Migrate `Prepare` to `Deploy` in `deployer` iface
This commit reworks the `deployer` internal interface to abstract
the whole mechanism of deploying the inputs to `kraft cloud deploy`
to the individual context.  This results in a new `Deploy` method
as opposed to a `Prepare` method which performs the entire action.
Instead of returning the package, which could be in fact an
intermediate step, the output is the (list) of instance(s).  This
greater abstraction provides future implementations to handle
different contexts that do not yield intermediate package references
which the `Prepare` method previously assumed.

The resulting instance(s) are properly output as the result from
the new `Deploy` method interface.
